Job_Search/requirements.txt
beautifulsoup4
python-dotenv
requests
openai